+++ vedo/plotter.py
@@ -2984,6 +2984,9 @@ class Plotter:
 
             elif isinstance(a, vtk.vtkImageData):
                 scanned_acts.append(vedo.Volume(a).actor)
+            
+            elif isinstance(a, vedo.RectilinearGrid):
+                scanned_acts.append(a.actor)
 
             elif isinstance(a, vtk.vtkLight):
                 self.renderer.AddLight(a)